+++ OpenMarket/OpenMarket/Model/Decoder.swift
@@ -5,13 +5,13 @@
 //  Created by 고은 on 2022/01/07.
 //
 
-import UIKit
+import Foundation
 
 struct Decoder {
-    let decoder = JSONDecoder()
-    
-    func parsePageJSON (data: Data) -> ProductList? {
+    func parseJSON (data: Data) -> ProductList? {
         do {
+            let decoder = JSONDecoder()
+            
             let pageJSON: Data = data
             let decodedPageJSON = try decoder.decode(ProductList.self, from: pageJSON)
             return decodedPageJSON
