+++ OpenMarket/OpenMarket/Model/NetworkModel/Common/ProductDetail.swift
